Switch to a more explicitly asynchronous model
This introduces a second command queue for submitting commands
to a WebView. It also switches to a model where there's a 1:1
correspondence between web view clients and web view requests.
This allows us to properly track any and all errors that might occur
for a given request. This seems to correctly implement theme setting
and restoration, although there are some minor visual issues as the
web view is displaying content just before the reading position and
theme are restored. This will be addressed in the next commits.

Affects: https://jira.nypl.org/browse/SIMPLY-2571