Implement full theme handling
This removes the internal command types in favour of a single command
submission type. It adjusts the controller to publish events when a
command will take long enough to execute that a progress bar should
ideally be shown. It also implements the various calls to views to
set color schemes.

Affects: https://jira.nypl.org/browse/SIMPLY-2571