Introduce a navigation graph
This introduces a strongly-typed navigation graph that is built from
a Publication and should allow for efficient and correct navigation
actions. Currently, the controller does not use the new graph, but
a test suite has been introduced in order to start checking that the
navigation graph gives the expected results. This necessitated moving
the sample epub files into the test resources.

Affects: https://jira.nypl.org/browse/SMA-192