Extend the graph with navigation targets
This adds the concept of navigation targets, which is essentially
a navigation node with an extra fragment. This is desirable because
it's not required that all identifiers inside pages/chapters be
declared in the TOC or spine. This means that, without taking into
account the possibly supplemental nature of fragments, a request
for:

  /OEBPS/chapter0.xhtml#figure-1

Would fail if the TOC or spine only advertised:

  /OEBPS/chapter0.xhtml

... Which many do. The new code takes a request such as the first
one, and searches for an exact match in the TOC, reading order, or
resources, and if no match is found, it removes the fragment and
performs the search again. If _this_ time a match is found, the
removed fragment is recorded as an extra and saved in the resulting
navigation target.

Currently, a bug in the R2 Javascript prevents navigating properly
to fragments, but this will be addressed in a subsequent commit.

Affects: https://jira.nypl.org/browse/SMA-192