Change the way page numbers are computed. There are now two options, both based on byte count rather than screens: - per chapter numbering - per book numbering
Fix the way non-linear resources are handled. The table of contents is not a resource collection per se.
It contains only shorcuts to items (maybe with fragments) in readingOrder or resources collections.
It might skip some readingOrder items, though swipping left or right should not ignore them.

The navigation graph is moved to vanilla and kept internal.
Retrieving resource titles from the TOC is done consistently and more rigourously.