Upgrade to R2 2.1.0 (#8)
This upgrades R2 to 2.1.0 in order to use a consistent set of R2
libraries and avoid a binary compatibility issue that prevented books
from opening. Currently, the served JS files have been merged into a
single readium-reflowable.js file that is served for all reflowable
EPUBs. This file has been duplicated and named readium-fixed.js,
and that file is served for fixed-layout EPUBs. This duplication of
source files is less than ideal, and hopefully a later R2 release
will give us a way to serve the same file for both requests.