Fix some race conditions
This changes the API to require clients to pass in an initial list of
bookmarks. This should ensure that it's not possible for the reader to
switch to a page _before_ having been informed that there's already a
bookmark for a later page. This also removes some deprecated events,
and replaces the broken bookmark creation with a much simpler system.

Additionally, view events are now published from a
BehaviorSubject. This allows for activities to avoid mild race
conditions when resuming.

Affects: https://ebce-lyrasis.atlassian.net/browse/PP-1120
Affects: https://ebce-lyrasis.atlassian.net/browse/PP-1119